pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fomorian"
version = "1.0.0"
description = "Attack scenario generator for Wazuh SIEM testing"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "PookaSec"}
]
keywords = [
    "security",
    "siem",
    "wazuh",
    "attack-simulation",
    "purple-team",
    "detection-engineering"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
docker = [
    "docker>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/pookasecteam/fomorian-community"
Documentation = "https://github.com/pookasecteam/fomorian-community#readme"
Repository = "https://github.com/pookasecteam/fomorian-community"
Issues = "https://github.com/pookasecteam/fomorian-community/issues"

[project.scripts]
fomorian = "generator.cli:cli"

[tool.setuptools]
packages = ["generator", "generator.config", "generator.builder", "generator.templating", "generator.output", "generator.injectors", "generator.wizard", "generator.wizard.steps", "generator.wazuh", "generator.random", "generator.preflight", "generator.preflight.checks"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
